scraper-pipeline/pipeline/requirements.txt
feedparser
bs4
requests
openai
textblob
nltk
python-dotenv
psycopg2-binary